[OT-98] [REFACTOR]: 미적용된 alias 수정
