+++ src/domains/mypage/components/recenthistory/RecentContentBox.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import RecentContentList from "@/components/mypage/recenthistory/RecentContentList";
+import RecentContentList from "@/domains/mypage/components/recenthistory/RecentContentList";
 import { mockRecentData } from "@/mocks/mockRecent";
 
 export default function RecentContentBox() {
